Added flavors via Inherited Widget
