+++ test/widget_test.dart
@@ -7,11 +7,13 @@
 
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:carriage_rider/main.dart';
+import 'package:carriage_rider/main_common.dart';
 
 void main() {
   testWidgets('MyApp can be loaded.', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(MyApp());
+
+    expect(1 + 1, 2);
   });
 }
